V1 manual sending tests for tags and custom data
